+++ CHANGES.rst
@@ -53,7 +53,7 @@ API changes
     ``BasicPSFPhotometry``, ``IterativelySubtractedPSFPhotometry`` and
     ``DAOPhotPSFPhotometry``. [#745]
 
-  - Added method ``_compute_local_model`` method to ``GriddedPSFModel``.
+  - Added method ``_compute_local_model`` to ``GriddedPSFModel``. [#1012]
 
 - ``photutils.segmentation``